--- a/lib/uri/idna/uts46/mapping.rb
+++ b/lib/uri/idna/uts46/mapping.rb
@@ -9,31 +9,35 @@ module URI
       module Mapping
         class << self
           def call(domain_name, transitional_processing: false, use_std3_ascii_rules: true)
-            output = +""
+            output = []
             domain_name.each_codepoint do |codepoint|
               _, status, replacement = status(codepoint)
               case status
               when "V", "X" # valid, disallowed
-                output << codepoint.chr(Encoding::UTF_8)
+                output << codepoint
               when "M" # mapped
-                output << if transitional_processing && codepoint == 7838
-                            "ss"
+                output += if transitional_processing && codepoint == 7838
+                            [115, 115]
                           else
                             replacement
                           end
               when "D" # deviation
-                output << (transitional_processing ? replacement : codepoint.chr(Encoding::UTF_8))
+                if transitional_processing
+                  output += replacement
+                else
+                  output << codepoint
+                end
               when "3" # disallowed_STD3_valid, disallowed_STD3_mapped
-                output << if use_std3_ascii_rules
-                            codepoint.chr(Encoding::UTF_8)
-                          else
-                            (replacement || codepoint.chr(Encoding::UTF_8))
-                          end
+                if use_std3_ascii_rules || !replacement
+                  output << codepoint
+                else
+                  output += replacement
+                end
               when "I" # ignored
                 next
               end
             end
-            output.unicode_normalize(:nfc)
+            output.pack("U*").unicode_normalize!(:nfc)
           end
 
           def validate_status(label, cp, pos, transitional_processing:, use_std3_ascii_rules:)
@@ -46,12 +50,9 @@ module URI
           end
 
           def status(codepoint)
-            index =
-              if codepoint < 256
-                codepoint
-              else
-                (UTS46_DATA.bsearch_index { |x| x[0] > codepoint } || UTS46_DATA.length) - 1
-              end
+            return UTS46_DATA[codepoint] if codepoint < 256
+
+            index = (UTS46_DATA.bsearch_index { |x| x[0] > codepoint } || UTS46_DATA.length) - 1
             UTS46_DATA[index] || []
           end
         end