… lib/uri/idna/data/uts46.rb