Various performance and memory optimizations.

Benchmarks:

```ruby
value = "google.com"
uni = "fiᆵリ宠퐱卄.com"
puny = "xn--fi-w1k207vk59a3qk9w9r.com"
N = 10_000

report = MemoryProfiler.report do
  N.times {
    URI::IDNA.to_unicode(URI::IDNA.to_ascii(value))
  }
  N.times {
    URI::IDNA.to_unicode(URI::IDNA.to_ascii(uni))
  }
  N.times {
    URI::IDNA.to_unicode(URI::IDNA.to_ascii(puny))
  }
end

report.pretty_print
```

```
BEFORE:
Total allocated: 208940164 bytes (3613842 objects)
Total retained:  643926 bytes (8102 objects)

AFTER:
Total allocated: 162245764 bytes (2176480 objects)
Total retained:  749526 bytes (10740 objects)
```

516 success cases from `IdnaTestV2` test suite:

```
BEFORE:
Calculating -------------------------------------
          to_unicode     84.340  (± 1.2%) i/s -    424.000  in   5.028420s
            to_ascii     66.525  (± 6.0%) i/s -    336.000  in   5.070995s
         to_ascii tr     66.506  (± 4.5%) i/s -    335.000  in   5.055958s

AFTER:
Calculating -------------------------------------
          to_unicode     59.628  (± 0.0%) i/s -    300.000  in   5.031440s
            to_ascii     50.837  (± 0.0%) i/s -    255.000  in   5.016066s
         to_ascii tr     48.980  (± 8.2%) i/s -    245.000  in   5.040046s